requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3

# Google Sheets integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0

# Data visualization
plotly==5.17.0

# Environment and configuration
python-dotenv==1.0.0

# Data processing and export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Additional utilities (optional)
requests==2.31.0
urllib3==2.0.7

# Development and testing (optional)
pytest==7.4.3
pytest-streamlit==1.0.0

# Performance monitoring (optional)
psutil==5.9.6

# Date and time utilities
python-dateutil==2.8.2